backend/requirements.txt
fastapi==0.115.12
python-multipart==0.0.20
pillow==11.2.1
pydantic==2.11.3
pytest==8.3.5
httpx==0.28.1
pytest-asyncio==0.26.0 
python-dotenv==1.1.0
boto3==1.38.2